% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_homogeneity_by_group}
\alias{check_homogeneity_by_group}
\title{Check homogeneity of variances across groups}
\usage{
check_homogeneity_by_group(data, x, y)
}
\arguments{
\item{data}{the data frame}

\item{x}{the grouping variable (column name as string)}

\item{y}{the dependent variable (column name as string)}
}
\value{
TRUE if Levene's test is non-significant (p >= .05), FALSE otherwise
}
\description{
Check homogeneity of variances across groups
}
