% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_normality_by_group}
\alias{check_normality_by_group}
\title{Check normality for groups}
\usage{
check_normality_by_group(data, x, y)
}
\arguments{
\item{data}{the data frame}

\item{x}{the x column}

\item{y}{the y column}
}
\value{
TRUE if all groups are normal, FALSE otherwise
}
\description{
Check normality for groups
}
