% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_betas_from_neale.R
\name{get_betas_from_neale}
\alias{get_betas_from_neale}
\title{Load and prune SNP effect sizes from Neale lab files}
\arguments{
\item{neale_filename}{path to the GWAS summary statistics file}

\item{variants_filename}{path to the variants information file}

\item{threshold}{p-value threshold to filter SNPs. Default is 5*10^-8}

\item{prune_distance}{distance (in base pairs) to determine independence of
SNPs. Default is 5*10^5 (500 kb)}
}
\value{
Returns a \code{data.frame} containing the RSIDs and beta effect
estimates of the pruned SNPs, with column names 'rsid' and 'beta'
}
\description{
Load and prune SNP effect sizes from Neale lab files
}
\examples{
  library( GxE )
  download.file( paste0( 'https://www.dropbox.com/s/gulqmkaighh8w3b/',
                         '21001_irnt.gwas.imputed_v3.both_sexes.tsv.bgz?raw=1' ),
                 '21001_irnt.gwas.imputed_v3.both_sexes.tsv.gz',
                 method = 'libcurl' )
  download.file( paste0( 'https://www.dropbox.com/s/puxks683vb0omeg/',
                         'variants.tsv.bgz?raw=1' ),
                 'variants.tsv.gz',
                 method = 'libcurl' )
  snps = get_betas_from_neale( neale_filename = '21001_irnt.gwas.imputed_v3.both_sexes.tsv.gz',
                               variants_filename = 'variants.tsv.gz' )

}
