% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize}
\alias{normalize}
\title{This function normalizes the values in a vector to the range [new_min, new_max]
based on their original range [old_min, old_max].}
\usage{
normalize(x_vector, old_min, old_max, new_min, new_max)
}
\arguments{
\item{x_vector}{A numeric vector that you want to normalize.}

\item{old_min}{The minimum value in the original scale of the data.}

\item{old_max}{The maximum value in the original scale of the data.}

\item{new_min}{The minimum value in the new scale to which you want to normalize the data.}

\item{new_max}{The maximum value in the new scale to which you want to normalize the data.}
}
\value{
A numeric vector with the normalized values.
}
\description{
This function normalizes the values in a vector to the range [new_min, new_max]
based on their original range [old_min, old_max].
}
\examples{
normalize(c(1, 2, 3, 4, 5), 1, 5, 0, 1)
}
