% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{not_empty}
\alias{not_empty}
\title{Ensure input is not empty}
\usage{
not_empty(x, msg = "Input must not be empty.")
}
\arguments{
\item{x}{The object to check}

\item{msg}{The error message to display}
}
\value{
Invisible TRUE if valid.
}
\description{
Stops execution if x is NULL, empty, or contains only NAs.
}
