% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{reportggstatsplot}
\alias{reportggstatsplot}
\title{Report statistical details for ggstatsplot.}
\usage{
reportggstatsplot(
  p,
  iv = "independent",
  dv = "Testdependentvariable",
  write_to_clipboard = FALSE
)
}
\arguments{
\item{p}{the object returned by ggwithinstats or ggbetweenstats}

\item{iv}{the independent variable}

\item{dv}{the dependent variable}

\item{write_to_clipboard}{whether to write to the clipboard}
}
\value{
A message describing the statistical results.
}
\description{
Report statistical details for ggstatsplot.
}
\examples{
\donttest{
library(ggstatsplot)
library(dplyr)

# Generate a plot
plt <- ggbetweenstats(mtcars, am, mpg)

reportggstatsplot(plt, iv = "am", dv = "mpg")
}
}
