% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rFromNPAV}
\alias{rFromNPAV}
\title{Calculation based on Rosenthal's formula (1994). N stands for the \emph{number of measurements}.
Necessary command:}
\usage{
rFromNPAV(pvalue, N)
}
\arguments{
\item{pvalue}{p value}

\item{N}{number of measurements in the experiment}
}
\value{
Invisibly returns a list with components:
\itemize{
\item \code{r}: effect size as a numeric scalar.
\item \code{z}: corresponding z-statistic.
\item \code{text}: LaTeX-formatted character string that is also sent
to the console.
}
}
\description{
Calculation based on Rosenthal's formula (1994). N stands for the \emph{number of measurements}.
Necessary command:
}
\examples{
rFromNPAV(0.02, N = 180)
}
